requirements.txt
bilibili-api-python>=18.0.0
requests>=2.28.0
Pillow>=10.0.0
pywin32>=305
curl-cffi>=0.7.0
